skip casting of literal ints

+++ src/fblas.rs
@@ -2,7 +2,7 @@
 #[no_mangle]
 pub unsafe extern "C" fn CINTdset0(mut n: i32, mut x: *mut f64) {
     let mut i: i32 = 0;
-    i = 0 as i32;
+    i = 0;
     while i < n {
         *x.offset(i as isize) = 0 as f64;
         i += 1;
@@ -18,7 +18,7 @@ pub unsafe extern "C" fn CINTdaxpy2v(
     mut v: *mut f64,
 ) {
     let mut i: i32 = 0;
-    i = 0 as i32;
+    i = 0;
     while i < n {
         *v.offset(i as isize) = a * *x.offset(i as isize) + *y.offset(i as isize);
         i += 1;
@@ -36,14 +36,14 @@ pub unsafe extern "C" fn CINTdmat_transpose(
     let mut i: i32 = 0;
     let mut j: i32 = 0;
     let _k: i32 = 0;
-    j = 0 as i32;
+    j = 0;
     while j < n - 3 as i32 {
-        i = 0 as i32;
+        i = 0;
         while i < m {
             *a_t
                 .offset(
-                    ((j + 0 as i32) * m + i) as isize,
-                ) = *a.offset((i * n + j + 0 as i32) as isize);
+                    ((j + 0) * m + i) as isize,
+                ) = *a.offset((i * n + j + 0) as isize);
             *a_t
                 .offset(
                     ((j + 1 as i32) * m + i) as isize,
@@ -63,7 +63,7 @@ pub unsafe extern "C" fn CINTdmat_transpose(
     }
     match n - j {
         1 => {
-            i = 0 as i32;
+            i = 0;
             while i < m {
                 *a_t.offset((j * m + i) as isize) = *a.offset((i * n + j) as isize);
                 i += 1;
@@ -71,12 +71,12 @@ pub unsafe extern "C" fn CINTdmat_transpose(
             }
         }
         2 => {
-            i = 0 as i32;
+            i = 0;
             while i < m {
                 *a_t
                     .offset(
-                        ((j + 0 as i32) * m + i) as isize,
-                    ) = *a.offset((i * n + j + 0 as i32) as isize);
+                        ((j + 0) * m + i) as isize,
+                    ) = *a.offset((i * n + j + 0) as isize);
                 *a_t
                     .offset(
                         ((j + 1 as i32) * m + i) as isize,
@@ -86,12 +86,12 @@ pub unsafe extern "C" fn CINTdmat_transpose(
             }
         }
         3 => {
-            i = 0 as i32;
+            i = 0;
             while i < m {
                 *a_t
                     .offset(
-                        ((j + 0 as i32) * m + i) as isize,
-                    ) = *a.offset((i * n + j + 0 as i32) as isize);
+                        ((j + 0) * m + i) as isize,
+                    ) = *a.offset((i * n + j + 0) as isize);
                 *a_t
                     .offset(
                         ((j + 1 as i32) * m + i) as isize,
@@ -117,12 +117,12 @@ pub unsafe extern "C" fn CINTdplus_transpose(
     let mut i: i32 = 0;
     let mut j: i32 = 0;
     let _k: i32 = 0;
-    j = 0 as i32;
+    j = 0;
     while j < n - 3 as i32 {
-        i = 0 as i32;
+        i = 0;
         while i < m {
-            *a_t.offset(((j + 0 as i32) * m + i) as isize)
-                += *a.offset((i * n + j + 0 as i32) as isize);
+            *a_t.offset(((j + 0) * m + i) as isize)
+                += *a.offset((i * n + j + 0) as isize);
             *a_t.offset(((j + 1 as i32) * m + i) as isize)
                 += *a.offset((i * n + j + 1 as i32) as isize);
             *a_t.offset(((j + 2 as i32) * m + i) as isize)
@@ -136,7 +136,7 @@ pub unsafe extern "C" fn CINTdplus_transpose(
     }
     match n - j {
         1 => {
-            i = 0 as i32;
+            i = 0;
             while i < m {
                 *a_t.offset((j * m + i) as isize) += *a.offset((i * n + j) as isize);
                 i += 1;
@@ -144,10 +144,10 @@ pub unsafe extern "C" fn CINTdplus_transpose(
             }
         }
         2 => {
-            i = 0 as i32;
+            i = 0;
             while i < m {
-                *a_t.offset(((j + 0 as i32) * m + i) as isize)
-                    += *a.offset((i * n + j + 0 as i32) as isize);
+                *a_t.offset(((j + 0) * m + i) as isize)
+                    += *a.offset((i * n + j + 0) as isize);
                 *a_t.offset(((j + 1 as i32) * m + i) as isize)
                     += *a.offset((i * n + j + 1 as i32) as isize);
                 i += 1;
@@ -155,10 +155,10 @@ pub unsafe extern "C" fn CINTdplus_transpose(
             }
         }
         3 => {
-            i = 0 as i32;
+            i = 0;
             while i < m {
-                *a_t.offset(((j + 0 as i32) * m + i) as isize)
-                    += *a.offset((i * n + j + 0 as i32) as isize);
+                *a_t.offset(((j + 0) * m + i) as isize)
+                    += *a.offset((i * n + j + 0) as isize);
                 *a_t.offset(((j + 1 as i32) * m + i) as isize)
                     += *a.offset((i * n + j + 1 as i32) as isize);
                 *a_t.offset(((j + 2 as i32) * m + i) as isize)
@@ -184,18 +184,18 @@ pub unsafe extern "C" fn CINTdgemm_NN1(
     let mut j: i32 = 0;
     let mut kp: i32 = 0;
     let mut bi: f64 = 0.;
-    j = 0 as i32;
+    j = 0;
     while j < n {
-        i = 0 as i32;
+        i = 0;
         while i < m {
             *c.offset((i + ldc * j) as isize) = 0 as f64;
             i += 1;
             i;
         }
-        kp = 0 as i32;
+        kp = 0;
         while kp < k {
             bi = *b.offset((kp + k * j) as isize);
-            i = 0 as i32;
+            i = 0;
             while i < m {
                 *c.offset((i + ldc * j) as isize)
                     += *a.offset((i + m * kp) as isize) * bi;
@@ -233,12 +233,12 @@ pub unsafe extern "C" fn CINTdgemm_TN(
     let mut j: i32 = 0;
     let mut kp: i32 = 0;
     let mut ci: f64 = 0.;
-    j = 0 as i32;
+    j = 0;
     while j < n {
-        i = 0 as i32;
+        i = 0;
         while i < m {
             ci = 0 as f64;
-            kp = 0 as i32;
+            kp = 0;
             while kp < k {
                 ci
                     += *a.offset((kp + k * i) as isize)
@@ -267,18 +267,18 @@ pub unsafe extern "C" fn CINTdgemm_NT(
     let mut j: i32 = 0;
     let mut kp: i32 = 0;
     let mut bi: f64 = 0.;
-    j = 0 as i32;
+    j = 0;
     while j < n {
-        i = 0 as i32;
+        i = 0;
         while i < m {
             *c.offset((i + m * j) as isize) = 0 as f64;
             i += 1;
             i;
         }
-        kp = 0 as i32;
+        kp = 0;
         while kp < k {
             bi = *b.offset((j + n * kp) as isize);
-            i = 0 as i32;
+            i = 0;
             while i < m {
                 *c.offset((i + m * j) as isize) += *a.offset((i + m * kp) as isize) * bi;
                 i += 1;
